+++ test/api/estimate-coin-convert.test.js
@@ -73,9 +73,9 @@ describe('EstimateCoinSell', () => {
             expect.assertions(2);
 
             return minterGate.estimateCoinSell({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToSell: 1,
-                coinIdToBuy: 1,
+                coinToBuy: 1,
             })
                 .then((estimateResult) => {
                     expect(Number(estimateResult.will_get)).toBeGreaterThan(0);
@@ -89,9 +89,9 @@ describe('EstimateCoinSell', () => {
         test('should fail gate', () => {
             expect.assertions(1);
             return minterGate.estimateCoinSell({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToSell: 1,
-                coinIdToBuy: 0,
+                coinToBuy: 0,
             })
                 .catch((error) => {
                     // logError(error);
@@ -103,9 +103,9 @@ describe('EstimateCoinSell', () => {
             expect.assertions(2);
 
             return minterNode.estimateCoinSell({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToSell: 1,
-                coinIdToBuy: 1,
+                coinToBuy: 1,
             })
                 .then((estimateResult) => {
                     expect(Number(estimateResult.will_get)).toBeGreaterThan(0);
@@ -119,9 +119,129 @@ describe('EstimateCoinSell', () => {
         test('should fail node', () => {
             expect.assertions(1);
             return minterNode.estimateCoinSell({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToSell: 1,
-                coinIdToBuy: 0,
+                coinToBuy: 0,
+            })
+                .catch((error) => {
+                    expect(error.response.data.error.message.length).toBeGreaterThan(0);
+                });
+        }, 30000);
+    });
+});
+
+describe('EstimateCoinSellAll', () => {
+    describe('symbol', () => {
+        test('should work gate', () => {
+            expect.assertions(1);
+
+            return minterGate.estimateCoinSellAll({
+                coinToSell: 'MNT',
+                valueToSell: 1,
+                coinToBuy: ENV_DATA.customCoin,
+            })
+                .then((estimateResult) => {
+                    expect(Number(estimateResult.will_get)).toBeGreaterThan(0);
+                })
+                .catch((error) => {
+                    logError(error);
+                });
+        }, 30000);
+
+        test('should fail gate', () => {
+            expect.assertions(1);
+            return minterGate.estimateCoinSellAll({
+                coinToSell: 'MNT',
+                valueToSell: 1,
+                coinToBuy: 'MNT',
+            })
+                .catch((error) => {
+                    logError(error);
+                    expect(error.response.data.error.message.length).toBeGreaterThan(0);
+                });
+        }, 30000);
+
+        test('should work node', () => {
+            expect.assertions(1);
+
+            return minterNode.estimateCoinSellAll({
+                coinToSell: 'MNT',
+                valueToSell: 1,
+                coinToBuy: ENV_DATA.customCoin,
+            })
+                .then((estimateResult) => {
+                    expect(Number(estimateResult.will_get)).toBeGreaterThan(0);
+                })
+                .catch((error) => {
+                    logError(error);
+                });
+        }, 30000);
+
+        test('should fail node', () => {
+            expect.assertions(1);
+            return minterNode.estimateCoinSellAll({
+                coinToSell: 'MNT',
+                valueToSell: 1,
+                coinToBuy: 'MNT',
+            })
+                .catch((error) => {
+                    expect(error.response.data.error.message.length).toBeGreaterThan(0);
+                });
+        }, 30000);
+    });
+
+    describe('id', () => {
+        test('should work gate', () => {
+            expect.assertions(1);
+
+            return minterGate.estimateCoinSellAll({
+                coinToSell: 0,
+                valueToSell: 1,
+                coinToBuy: 1,
+            })
+                .then((estimateResult) => {
+                    expect(Number(estimateResult.will_get)).toBeGreaterThan(0);
+                })
+                .catch((error) => {
+                    logError(error);
+                });
+        }, 30000);
+
+        test('should fail gate', () => {
+            expect.assertions(1);
+            return minterGate.estimateCoinSellAll({
+                coinToSell: 0,
+                valueToSell: 1,
+                coinToBuy: 0,
+            })
+                .catch((error) => {
+                    // logError(error);
+                    expect(error.response.data.error.message.length).toBeGreaterThan(0);
+                });
+        }, 30000);
+
+        test('should work node', () => {
+            expect.assertions(1);
+
+            return minterNode.estimateCoinSellAll({
+                coinToSell: 0,
+                valueToSell: 1,
+                coinToBuy: 1,
+            })
+                .then((estimateResult) => {
+                    expect(Number(estimateResult.will_get)).toBeGreaterThan(0);
+                })
+                .catch((error) => {
+                    logError(error);
+                });
+        }, 30000);
+
+        test('should fail node', () => {
+            expect.assertions(1);
+            return minterNode.estimateCoinSellAll({
+                coinToSell: 0,
+                valueToSell: 1,
+                coinToBuy: 0,
             })
                 .catch((error) => {
                     expect(error.response.data.error.message.length).toBeGreaterThan(0);
@@ -197,9 +317,9 @@ describe('EstimateCoinBuy', () => {
             expect.assertions(2);
 
             return minterGate.estimateCoinBuy({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToBuy: 1,
-                coinIdToBuy: 1,
+                coinToBuy: 1,
             })
                 .then((estimateResult) => {
                     expect(Number(estimateResult.will_pay)).toBeGreaterThan(0);
@@ -213,9 +333,9 @@ describe('EstimateCoinBuy', () => {
         test('should fail gate', () => {
             expect.assertions(1);
             return minterGate.estimateCoinBuy({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToBuy: 1,
-                coinIdToBuy: 0,
+                coinToBuy: 0,
             })
                 .catch((error) => {
                     // console.log(error.response.data);
@@ -227,9 +347,9 @@ describe('EstimateCoinBuy', () => {
             expect.assertions(2);
 
             return minterNode.estimateCoinBuy({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToBuy: 1,
-                coinIdToBuy: 1,
+                coinToBuy: 1,
             })
                 .then((estimateResult) => {
                     expect(Number(estimateResult.will_pay)).toBeGreaterThan(0);
@@ -243,9 +363,9 @@ describe('EstimateCoinBuy', () => {
         test('should fail node', () => {
             expect.assertions(1);
             return minterNode.estimateCoinBuy({
-                coinIdToSell: 0,
+                coinToSell: 0,
                 valueToBuy: 1,
-                coinIdToBuy: 0,
+                coinToBuy: 0,
             })
                 .catch((error) => {
                     expect(error.response.data.error.message.length).toBeGreaterThan(0);